Safari-611.1.10/LayoutTests/fast/dom/Window/webkitConvertPointUpdateLayout-expected.txt

This test exercises the webkitConvertPointFromNodeToPage() function after changing the transform.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS x is 60
PASS y is 70
PASS successfullyParsed is true

TEST COMPLETE
